Resources in CloudFormation to deploy the application.
Mails are received via SES, stored in S3, and re-uploaded to an IMAP server
in a Lambda function.

Some configuration is required in SES, which is mentioned in the example
template.

The template is written with troposphere, and the stack is managed with
sceptre. All libraries can be installed with `pip install -r requirements.txt`.

The code of the handler in the Lambda function is not implemented.